feat: remove basics module
