+++ lib/merge_ruby_client.rb
@@ -9,7 +9,7 @@ require_relative "merge_ruby_client/hris/client"
 require_relative "merge_ruby_client/filestorage/client"
 require_relative "merge_ruby_client/ticketing/client"
 require_relative "merge_ruby_client/accounting/client"
-require_relative "merge_ruby_client/basics/client"
+require_relative "merge_ruby_client/integrations/client"
 
 module Merge
   class Client
@@ -25,8 +25,8 @@ module Merge
     attr_reader :ticketing
     # @return [Merge::Accounting::Client]
     attr_reader :accounting
-    # @return [Merge::Basics::Client]
-    attr_reader :basics
+    # @return [Merge::IntegrationClient]
+    attr_reader :integrations
 
     # @param base_url [String]
     # @param environment [Merge::Environment]
@@ -51,7 +51,7 @@ module Merge
       @filestorage = Merge::Filestorage::Client.new(request_client: @request_client)
       @ticketing = Merge::Ticketing::Client.new(request_client: @request_client)
       @accounting = Merge::Accounting::Client.new(request_client: @request_client)
-      @basics = Merge::Basics::Client.new(request_client: @request_client)
+      @integrations = Merge::IntegrationClient.new(request_client: @request_client)
     end
   end
 
@@ -68,8 +68,8 @@ module Merge
     attr_reader :ticketing
     # @return [Merge::Accounting::AsyncClient]
     attr_reader :accounting
-    # @return [Merge::Basics::AsyncClient]
-    attr_reader :basics
+    # @return [Merge::IntegrationClient]
+    attr_reader :integrations
 
     # @param base_url [String]
     # @param environment [Merge::Environment]
@@ -94,7 +94,7 @@ module Merge
       @filestorage = Merge::Filestorage::AsyncClient.new(request_client: @async_request_client)
       @ticketing = Merge::Ticketing::AsyncClient.new(request_client: @async_request_client)
       @accounting = Merge::Accounting::AsyncClient.new(request_client: @async_request_client)
-      @basics = Merge::Basics::AsyncClient.new(request_client: @async_request_client)
+      @integrations = Merge::IntegrationClient.new(request_client: @async_request_client)
     end
   end
 end
